[components] box refactoring
